requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
python-multipart>=0.0.6
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
Pillow>=10.2.0

# PostgreSQL driver (uncomment when switching to PostgreSQL)
# psycopg2-binary>=2.9.9